+++ src/indexer/rmq_con.rs
@@ -1,11 +1,27 @@
 use std::collections::HashSet;
 
 use anyhow::Context;
+use chrono::Utc;
+use lapin::options::ConfirmSelectOptions;
 use lapin::tcp::{AMQPUriTcpExt, NativeTlsConnector};
 use lapin::uri::AMQPUri;
 use lapin::{Connection, ConnectionProperties};
+use rand::distributions::{Alphanumeric, DistString};
 
-pub async fn connect_to_rabbitmq(addr: &str) -> Result<Connection, anyhow::Error> {
+pub async fn setup_rabbitmq_connection(addr: &str) -> Result<lapin::Channel, anyhow::Error> {
+  let conn = connect_to_rabbitmq(addr).await?;
+  let channel = conn
+    .create_channel()
+    .await
+    .context("creates rmq connection channel")?;
+  channel
+    .confirm_select(ConfirmSelectOptions::default())
+    .await
+    .context("enable msg confirms")?;
+  Ok(channel)
+}
+
+async fn connect_to_rabbitmq(addr: &str) -> Result<Connection, anyhow::Error> {
   let uri = addr
     .parse::<AMQPUri>()
     .map_err(anyhow::Error::msg)
@@ -34,4 +50,15 @@ pub async fn connect_to_rabbitmq(addr: &str) -> Result<Connection, anyhow::Error
       .await
       .context("Failed to establish an unsecure AMQP connection")
   }
+}
+
+// TODO get pod name from k8s?
+pub fn generate_consumer_tag(prefix: &str) -> String {
+  let timestamp = Utc::now().format("%Y%m%d%H%M%S");
+  format!(
+    "{}-{}-{}",
+    prefix,
+    timestamp,
+    Alphanumeric.sample_string(&mut rand::thread_rng(), 16)
+  )
 }
